\name{jomo-package}
\alias{jomo-package}
\alias{jomo}
\docType{package}
\title{
A package for Multilevel Joint Modelling Multiple Imputation
}
\description{
Building on Schafer's pan and on standalone program REALCOM, jomo is a package for multilevel joint modelling multiple imputation.
Binary and categorical variables are handled through latent normal variables and algorithms for cluster specific covariance matrices are introduced. 
}
\details{
\tabular{ll}{
Package: \tab jomo\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2014-10-24\cr
License: \tab GPL-2\cr
}

}
\author{
Matteo Quartagno, London School of Hygiene and Tropical Medicine 
Professor James Carpenter, London School of Hygiene and Tropical Medicine, MRC Clinical Trials Unit at UCL

Maintainer:  Matteo Quartagno <matteo.quartagno@lshtm.ac.uk>
}
\references{
Carpenter J.R., Kenward M.G., (2013), Multiple Imputation and its Application. Chapters 3-5-9, Wiley, ISBN: 978-0-470-74052-1.

Yucel R.M., (2011), Random-covariances and mixed-effects models for imputing multivariate multilevel continuous data, Statistical Modelling, 11 (4), 351-370, DOI: 10.1177/1471082X100110040.

C libraries: pdflib, rnglib and wishart, John Burkardt, Dept. of Scientific Comupting, The Florida State University


}
