\name{jomo-package}
\alias{jomo-package}
\docType{package}
\title{
A package for Multilevel Joint Modelling Multiple Imputation
}
\description{
Similarly to Schafer's package pan, jomo is a package for multilevel joint modelling multiple imputation.
 Novel aspects of jomo are the possibility of handling binary and categorical data through latent normal variables and the option to use cluster-specific covariance matrices.
}
\details{
\tabular{ll}{
Package: \tab jomo\cr
Type: \tab Package\cr
Version: \tab 1.3-0\cr
Date: \tab 2015-12-14\cr
License: \tab GPL-2\cr
}

}
\author{
Matteo Quartagno, London School of Hygiene and Tropical Medicine 
Professor James Carpenter, London School of Hygiene and Tropical Medicine, MRC Clinical Trials Unit at UCL

Maintainer:  Matteo Quartagno <matteo.quartagno@lshtm.ac.uk>
}
\references{
Carpenter J.R., Kenward M.G., (2013), Multiple Imputation and its Application. Chapters 3-5-9, Wiley, ISBN: 978-0-470-74052-1.

Yucel R.M., (2011), Random-covariances and mixed-effects models for imputing multivariate multilevel continuous data, Statistical Modelling, 11 (4), 351-370, DOI: 10.1177/1471082X100110040.

C libraries: pdflib, rnglib and wishart, John Burkardt, Dept. of Scientific Comupting, The Florida State University


}
