\name{sldata}
\alias{sldata}
\docType{data}
\title{
A simulated single level dataset 
}
\description{
A simulated dataset to test single level functions, i.e. jomo1con, jomo1cat and jomo1mix. 
}
\usage{data(sldata)}
\format{
  A data frame with 300 observations on the following 4 variables.
  \describe{
    \item{\code{age}}{A numeric variable with age. Fully observed.}
    \item{\code{measure}}{A numeric variable with some measure of interest (unspecified). This is partially observed.}
    \item{\code{sex}}{A binary variable for gender indicator. Fully observed.}
    \item{\code{social}}{A 4-category variable with a social status indicator. This is partially observed.}

  }
}
\details{
These are not real data, they are simulated to illustrate the use of the main functions of the package.}

